ver 0.2 - feat. 3
노트 이동, 판정을 AudioSettings.dspTime 기반으로 변경
노트 이동에는 Time.time과 함께 보간

+++ Assets/Scripts/NoteForJudge.cs
@@ -18,7 +18,7 @@ public enum DirectionType
 
 public class NoteForJudge
 {
-    public float Time;
+    public double Time;
     public MoveType Type;
     public List<DirectionType> Directions;
 }
